requirements.txt
gradio>=4.0.0
transformers==4.40.0
torch>=2.0.0
Pillow>=9.0.0
numpy>=1.22.0
einops>=0.8.0
accelerate>=0.27.0
safetensors>=0.4.1
sentencepiece>=0.1.99
protobuf>=3.20.0
